pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neurospark-core"
version = "0.1.0"
description = "An autonomous agentic learning platform"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "NeuroSpark Team", email = "info@neurospark.example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.11.0",
    "psycopg2-binary>=2.9.6",
    "redis>=4.6.0",
    "grpclib>=0.4.5",
    "protobuf>=4.23.4",
    "llama-index>=0.8.0",
    "qdrant-client>=1.4.0",
    "minio>=7.1.15",
    "trafilatura>=1.6.0",
    "pdfminer.six>=20221105",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.1",
    "tenacity>=8.2.2",
    "pyyaml>=6.0",
    "rouge>=1.0.1",
    "bert-score>=0.3.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "flake8>=6.0.0",
    "pre-commit>=3.3.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "--cov=src --cov-report=term-missing"
